steering/gesten/requirements.txt
absl-py==1.4.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.8
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.14
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.29.2
shapely==2.0.1
six==1.16.0
sounddevice==0.4.6
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.43
ultralyticsplus==0.0.28
urllib3==1.26.16
Werkzeug==2.3.7
